Experimental changes to the modeling system.
Due to the fact that protected static variables don't automatically get
created and initialized for each subclass of the Model class, an new
experimental system is created. This system is similar to Doctrine
where the model itself won't interact with the Database. However,
different shortcut functions will still be provided to emulate Django
and GAE as close as possible.